+++ lib/puree/xml_extractor/research_output.rb
@@ -142,8 +142,8 @@ module Puree
         data = []
         xpath_result.each do |i|
           s = Puree::Model::ResearchOutputScopusMetric.new
-          s.value = i.xpath('value').text.strip
-          s.year = i.xpath('year').text.strip
+          s.value = i.xpath('value').text.strip.to_i
+          s.year = i.xpath('year').text.strip.to_i
           data << s
         end
         data