+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 2.5.1 2019-01-10
 ### Fixed
-- Research output - file URL.
+- Research output - file URL, scopus metric type
 
 ## 2.5.0 2019-01-04
 ### Added